Number of Zero-Filled Subarrays brute

+++ Arrays/2348.Number-of-Zero-Filled-Subarrays-brute.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long zeroFilledSubarray(vector<int>& nums) {
-        int count = 0;
+        long long count = 0;
         int n = nums.size();
         for(int i=0;i<n;i++){
             if(nums[i] !=0 ) continue;
